get rid of unused things; proper cellpose args
